+++ includes/field_splitter.h
@@ -19,4 +19,6 @@ t_token	*create_field_list(char **fields, t_token *original);
 t_token	*split_token(t_token *token, t_field_context *cntxt);
 void	append_token_list(t_token **dest, t_token *src);
 char	**ft_split_by_space(char const	*s);
+void	free_fields_array(char **fields);
+
 #endif
