+++ parsing/expander/fields/fields_list.c
@@ -1,34 +1,5 @@
 #include "../../../includes/minishell.h"
 
-int	contains_whitespace(char *str)
-{
-	int	i;
-
-	if (str == NULL)
-		return (0);
-	i = 0;
-	while (str[i] != '\0')
-	{
-		if (is_white_space(str[i]))
-			return (1);
-		i++;
-	}
-	return (0);
-}
-
-int	should_split_token(t_token *token, t_field_context *cntxt)
-{
-	if (token->type == D_QUOTE || token->type == S_QUOTE)
-		cntxt->was_quoted = 1;
-	cntxt->needs_splitting = 0;
-	if (token->type == WORD && cntxt->was_quoted == 0)
-	{
-		if (contains_whitespace(token->value))
-			cntxt->needs_splitting = 1;
-	}
-	return (cntxt->needs_splitting);
-}
-
 t_token	*create_field_list(char **fields, t_token *original)
 {
 	int		i;
@@ -60,10 +31,11 @@ t_token	*split_token(t_token *token, t_field_context *cntxt)
 	fields = ft_split_by_space(token->value);
 	if (fields == NULL || fields[0] == NULL)
 	{
-		/*printf("need to free?\n");*/
+		free_fields_array(fields);
 		return (create_token_node(token->value, token->type));
 	}
 	field_list = create_field_list(fields, token);
+	free_fields_array(fields);
 	return (field_list);
 }
 
@@ -72,7 +44,7 @@ void	append_token_list(t_token **dest, t_token *src)
 	t_token	*current;
 
 	if (src == NULL)
-		return (NULL);
+		return ;
 	if (*dest == NULL)
 	{
 		*dest = src;
