+++ parsing/expander/fields/fileds_splitter.c
@@ -4,6 +4,7 @@ t_token	*field_splitter(t_token *tokens)
 {
 	t_token			*current;
 	t_token			*split_tokens;
+	t_token			*new_node;
 	t_token			*new_list;
 	t_field_context	cntxt;
 
@@ -11,13 +12,19 @@ t_token	*field_splitter(t_token *tokens)
 	current = tokens;
 	while (current != NULL)
 	{
+		ft_memset(&cntxt, 0, sizeof(cntxt));
 		if (should_split_token(current, &cntxt))
 		{
 			split_tokens = split_token(current, &cntxt);
-			append_token_list(&new_list, split_token);
+			if (split_token != NULL)
+				append_token_list(&new_list, split_tokens);
 		}
 		else
-			add_token(&new_list, create_token_node(current->value, current->type));
+		{
+			new_node = create_token_node(current->value, current->type);
+			if (new_node != NULL)
+				add_token(&new_list, new_node);
+		}
 		current = current->next;
 	}
 	return (new_list);
